search-api: Make use of shared scalatra code
